+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes to this project are documented in this file. More detail (including info
 - AUTOTOOL-3704: Expose metadata property for pools and pool members
 - AUTOTOOL-3723: Add basic POST functionality for per-app declarations
 - AUTOTOOL-3876: Support for performing AS3 string expansion inside declarative WAF policy
+- AUTOTOOL-3866: Add support of route domain as a string for SOCKS profile
 
 ### Fixed
 - AUTOTOOL-3842: Updated list of services for Protocol_Inspection_Profile